refactor: remove dprint-cli-core crate
