--- crates/cli-core/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "dprint-cli-core"
-version = "0.12.3"
-authors = ["David Sherret <dsherret@gmail.com>"]
-edition = "2021"
-license = "MIT"
-repository = "https://github.com/dprint/dprint"
-description = "Core dprint CLI library for reuse."
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.58"
-crossterm = "0.24.0"
-parking_lot = "0.12.1"
-sha2 = "0.10.2"
-ureq = { version = "2.4.0", features = ["socks-proxy", "tls"] }
