--- crates/cli-core/src/lib.rs
@@ -1,6 +1,0 @@
-pub mod checksums;
-pub mod logging;
-pub mod terminal;
-mod url_utils;
-
-pub use url_utils::*;
