+++ crates/dprint/src/utils/checksums.rs
