+++ crates/dprint/src/utils/logging/confirm.rs
@@ -3,10 +3,11 @@ use anyhow::Result;
 use crossterm::event::Event;
 use crossterm::event::KeyCode;
 
-use crate::logging::Logger;
-use crate::logging::LoggerRefreshItemKind;
-use crate::logging::LoggerTextItem;
-use crate::terminal::read_terminal_event;
+use super::Logger;
+use super::LoggerRefreshItemKind;
+use super::LoggerTextItem;
+use crate::utils::terminal::get_terminal_width;
+use crate::utils::terminal::read_terminal_event;
 
 pub fn show_confirm(logger: &Logger, context_name: &str, prompt: &str, default_value: bool) -> Result<bool> {
   let result = loop {
@@ -43,7 +44,7 @@ pub fn show_confirm(logger: &Logger, context_name: &str, prompt: &str, default_v
   logger.log_text_items(
     &[LoggerTextItem::Text(format!("{} {}", prompt, if result { "Y" } else { "N" }))],
     context_name,
-    crate::terminal::get_terminal_width(),
+    get_terminal_width(),
   );
 
   Ok(result)
