+++ crates/dprint/src/utils/logging/log_action_with_progress.rs
