+++ crates/dprint/src/utils/mod.rs
@@ -1,4 +1,5 @@
 mod cached_downloader;
+mod checksums;
 mod error_count_logger;
 mod extract_zip;
 mod file_path_utils;
@@ -6,15 +7,20 @@ mod file_text;
 mod get_bytes_hash;
 mod get_difference;
 mod glob;
+mod logging;
 mod path_source;
 mod pretty_print_json_text;
 mod process;
 mod resolve_url_or_file_path;
 mod stdin_reader;
 mod table_text;
+mod terminal;
 mod update_checker;
+mod url;
 
+pub use self::url::*;
 pub use cached_downloader::*;
+pub use checksums::*;
 pub use error_count_logger::*;
 pub use extract_zip::*;
 pub use file_path_utils::*;
@@ -22,10 +28,12 @@ pub use file_text::*;
 pub use get_bytes_hash::*;
 pub use get_difference::*;
 pub use glob::*;
+pub use logging::*;
 pub use path_source::*;
 pub use pretty_print_json_text::*;
 pub use process::*;
 pub use resolve_url_or_file_path::*;
 pub use stdin_reader::*;
 pub use table_text::*;
+pub use terminal::*;
 pub use update_checker::*;
