+++ crates/dprint/src/utils/terminal.rs
